+++ Kmarket/src/main/resources/mappers/adminCs.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- 2023/02/13 // 심규영 // 관리자 고객센터 xml 생성 -->
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="kr.co.kmarket.dao.admin.AdminCsDAO">
+    <!-- create -->
+
+    <!-- read -->
+    <select id="selectCsArticles">
+        
+    </select>
+
+    <!-- upload -->
+
+
+    <!-- delete -->
+
+</mapper>
