+++ Kmarket/src/main/resources/templates/admin/product/list.html
@@ -4,6 +4,7 @@
 <head>
     <th:block th:replace="~{fragment/head_fragment::basicHead}"></th:block>
     <th:block th:replace="~{fragment/head_fragment::adminHead}"></th:block>
+    <script th:src="@{/js/admin/productList.js}"></script>
 </head>
 <body>
 <div id="admin-wrapper">
